% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patentsrd.R
\docType{data}
\name{patentsrd}
\alias{patentsrd}
\title{Patents and R&D}
\format{
A data frame with 346 observations and 25 variables:
\describe{
\item{\strong{cusip}}{Compustat's identifying number for the firm
(Committee on Uniform Security Identification Procedures number).}
\item{\strong{ardssic}}{A two-digit code for the applied R&D
industrial classification.}
\item{\strong{scisect}}{= 1 if firm is in the scientific sector.}
\item{\strong{logk}}{log of the book value of capital in 1972.}
\item{\strong{sumpat}}{sum of patents applied for between 1972-1979.}
\item{\strong{logr70 - logr79}}{logarithm of R&D spending during the year
(in 1972 dollars).}
\item{\strong{pat70 - pat79}}{number of patents applied
for during the year that were eventually granted.}
}
}
\source{
\url{http://cameron.econ.ucdavis.edu/mmabook/mmadata.html}
}
\usage{
data(patentsrd)
}
\description{
Panel data on patents and R&D expenditures. The sample includes
346 firms with five years of data from 1975 to 1979 used by Hall,
Griliches, and Hausman (1986).
}
\section{Section in Text}{

23.3 Nonlinear Panel Example: Patents and R&D, pp. 792-5
}

\examples{
dplyr::glimpse(patentsrd)
head(patentsrd)
}
\references{
Cameron, A. and Trivedi, P. (2005), "Microeconometrics: Methods and
Applications," Cambridge University Press, New York.

Hall, B., Griliches, Z. and Hausman J. (1986), "Patents and R and D:
Is There a Lag?," International Economic Review, 27,
issue 2, p. 265-83.
}
\keyword{datasets}
