% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randdata.R
\docType{data}
\name{randdata}
\alias{randdata}
\title{Health expenditures and insurance plans}
\format{
A data frame with 20,190 observations and 45 variables:
\describe{
\item{\strong{var}}{..}
\item{\strong{var}}{..}
...
}
}
\source{
\url{http://cameron.econ.ucdavis.edu/mmabook/mmadata.html}
}
\usage{
data(randdata)
}
\description{
Data from the RAND Health Insurance Experiment. The data extract comes
from Deb and Trivedi (2002).
Cameron and Trivedi (2005).
}
\section{Section in Text}{

16.6 Selection Models, pp. 553-6, 565
}

\references{
Cameron, A. and Trivedi, P. (2005), "Microeconometrics: Methods and
Applications," Cambridge University Press, New York.
}
\keyword{datasets}
