% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randhealth.R
\docType{data}
\name{randhealth}
\alias{randhealth}
\title{Health expenditures and insurance plans}
\format{
A data frame with 20,190 observations and 45 variables:
\describe{
\item{\strong{plan}}{health insurance plan number}
\item{\strong{site}}{one of six sites where experiment was conducted}
\item{\strong{coins}}{medical coinsurance}
\item{\strong{tookphys}}{took baseline physical}
\item{\strong{year}}{study year}
\item{\strong{zper}}{person id, leading digit is sit}
\item{\strong{black}}{= 1 if race of household head is black}
\item{\strong{income}}{income based on annual income}
\item{\strong{xage}}{age that year}
\item{\strong{female}}{= 1 if person is female}
\item{\strong{educdec}}{years of schooling of decision maker}
\item{\strong{time}}{time eligible during the year}
\item{\strong{outpdol}}{outpatient exp. excl. ment and}
\item{\strong{drugdol}}{drugs purchased, outpatient}
\item{\strong{suppdol}}{supplies purchased, outpatient}
\item{\strong{mentdol}}{psychotherapy exp., outpatient}
\item{\strong{inpdol}}{inpatient exp., facilities and md}
\item{\strong{meddol}}{annual medical expenditures in constant dollars, excluding dental and outpatient mental}
\item{\strong{totadm}}{number of hospital admissions}
\item{\strong{inpmis}}{missing any inpatient charges}
\item{\strong{mentvis}}{number psychotehrapy visits}
\item{\strong{mdvis}}{number face-to-face md visits}
\item{\strong{notmdvis}}{number face-to-face, not md visits}
\item{\strong{num}}{family size}
\item{\strong{mhi}}{mental health index, baseline}
\item{\strong{disea}}{number of chronic diseases}
\item{\strong{physlm}}{= 1 if person has physical limitation}
\item{\strong{ghindx}}{general health index, baseline}
\item{\strong{mdeoff}}{maximum expenditure offer}
\item{\strong{pioff}}{participation incentive}
\item{\strong{child}}{= 1 if age is less than 18}
\item{\strong{fchild}}{= 1 if female child}
\item{\strong{lfam}}{log of family size}
\item{\strong{lpi}}{log of annual participation incentive payment or 0 if no payment}
\item{\strong{idp}}{= 1 if individual deductible plan}
\item{\strong{logc}}{log(coinsurance + 1) where coinsurance rate is 0 to 100}
\item{\strong{fmde}}{log(max(medical deductible expenditure)) if idp=1 and mde>1, 0 otherwise}
\item{\strong{hlthg}}{= 1 if self-rated health is good}
\item{\strong{hlthf}}{= 1 if self-rated health is fair}
\item{\strong{hlthp}}{= 1 if self-rated health is poor, (omitted is excellent)}
\item{\strong{xghindx}}{ghi with imputation}
\item{\strong{linc}}{log of annual family income, usd}
\item{\strong{lnum}}{log of family size}
\item{\strong{lnmeddol}}{log of medical expenditures given meddol > 0; missing otherwise}
\item{\strong{binexp}}{= 1 if medical expenditures > 0}
}
}
\source{
\url{http://cameron.econ.ucdavis.edu/mmabook/mmadata.html}
}
\usage{
data(randhealth)
}
\description{
Data from the RAND Health Insurance Experiment. The data comes
from Deb and Trivedi (2002). It includes variables on the number of contacts
with a medical doctor, medical expenditures, demographics, health status, and insurance status.
Cameron and Trivedi (2005).
}
\section{Section in Text}{

16.6 Selection Models, pp. 553-6, 565
20.3 Count Example: Contacts with Medical Doctor, p.671
}

\references{
Cameron, A. and Trivedi, P. (2005), "Microeconometrics: Methods and
Applications," Cambridge University Press, New York.

Deb, P. and Trivedi, P.K. (2002), "The Structure of Demand for Health Care:
Latent Class versus Two-Part Models," Journal of Health Economics, 21, 601-625.

RAND Corporation. "RAND's Health Insurance Experiment ."
https://www.rand.org/health-care/projects/hie.html
}
\keyword{datasets}
